group_diagnostics_by_filename and MessageWithLocation to grouped
